Fix register command not working for first time users
Whenever a new user, who has never messaged Darknet
tries to register, the lookup fails to find their info
and errors out.
This fix skips orgbanlist checks for initial register
command and effectively allows new users to register.
This means that a user from banned org will be able to
register, but will not be able to interact with the bot.